@@ -2,7 +2,9 @@
   "name": "minio",
   "version": "7.0.34",
   "description": "S3 Compatible Cloud Storage client",
-  "main": "./dist/main/minio.js",
+  "source": "src/main/minio.ts",
+  "main": "./dist/cjs/minio.js",
+  "module": "dist/esm/index.js",
   "scripts": {
     "compile": "gulp compile",
     "test": "gulp test",
@@ -11,7 +13,23 @@
     "prepublishOnly": "gulp test",
     "functional": "gulp functionalTest",
     "browserify": "gulp browserify",
-    "prepare": "npm run compile"
+    "prepare": "husky install && npm run compile",
+    "format": "prettier -w .",
+    "format-check": "prettier --list-different .",
+    "lint-staged": "lint-staged"
+  },
+  "prettier": {
+    "printWidth": 120,
+    "endOfLine": "lf",
+    "semi": false
+  },
+  "lint-staged": {
+    "*.{js,cjs,mjs,ts,cts,mjs,json,yaml,yml}": [
+      "prettier --write"
+    ],
+    "src/**/*.{js,cjs,mjs,ts,cts,mjs}": [
+      "eslint --fix"
+    ]
   },
   "repository": {
     "type": "git",
@@ -53,23 +71,40 @@
   "devDependencies": {
     "@babel/core": "^7.12.10",
     "@babel/preset-env": "^7.12.10",
+    "@rollup/plugin-babel": "^6.0.3",
+    "@rollup/plugin-replace": "^5.0.2",
+    "@rollup/plugin-typescript": "^11.1.0",
+    "@types/lodash": "^4.14.192",
+    "@types/node": "^18.15.11",
+    "@typescript-eslint/eslint-plugin": "^5.57.1",
+    "@typescript-eslint/parser": "^5.57.1",
     "babelify": "^10.0.0",
     "browserify": "^16.5.2",
     "chai": "^4.2.0",
-    "eslint": "^8.9.0",
+    "dotenv": "^16.0.3",
+    "eslint": "^8.37.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-simple-import-sort": "^10.0.0",
     "gulp": "^4.0.2",
     "gulp-babel": "^8.0.0",
-    "gulp-eslint": "^4.0.2",
+    "gulp-eslint": "^6.0.0",
     "gulp-if": "^3.0.0",
     "gulp-mocha": "^8.0.0",
     "gulp-sourcemaps": "^2.6.5",
+    "husky": "^8.0.3",
+    "lint-staged": "^13.2.1",
     "mocha": "^9.2.0",
     "mocha-steps": "^1.1.0",
     "nock": "^13.2.2",
+    "prettier": "^2.8.7",
+    "rollup-plugin-node-externals": "^5.1.2",
     "source-map-support": "^0.5.13",
     "split-file": "^2.2.2",
-    "superagent": "^5.1.0",
-    "uuid": "^3.3.2"
+    "superagent": "^8.0.9",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.0.4",
+    "uuid": "^3.4.0"
   },
   "keywords": [
     "api",
