@@ -14,27 +14,27 @@
  * limitations under the License.
  */
 
- // Note: YOUR-ACCESSKEYID, YOUR-SECRETACCESSKEY, my-testfile, my-bucketname
- // and my-objectname are dummy values, please replace them with original values.
+// Note: YOUR-ACCESSKEYID, YOUR-SECRETACCESSKEY, my-testfile, my-bucketname
+// and my-objectname are dummy values, please replace them with original values.
 
-var Minio = require('minio')
-var Fs = require('fs')
+var Minio = require("minio")
+var Fs = require("fs")
 
 var s3Client = new Minio.Client({
-  endPoint: 's3.amazonaws.com',
-  accessKey: 'YOUR-ACCESSKEYID',
-  secretKey: 'YOUR-SECRETACCESSKEY'
+  endPoint: "s3.amazonaws.com",
+  accessKey: "YOUR-ACCESSKEYID",
+  secretKey: "YOUR-SECRETACCESSKEY",
 })
 
 var metaData = {
-  'Content-Type': 'application/octet-stream',
-  'X-Amz-Meta-Testing': 1234,
-  'example': 5678
+  "Content-Type": "application/octet-stream",
+  "X-Amz-Meta-Testing": 1234,
+  example: 5678,
 }
 
 // Put a file in bucket my-bucketname.
-var file = 'my-testfile'
-s3Client.fPutObject('my-bucketname', 'my-objectname', file, metaData, function(e) {
+var file = "my-testfile"
+s3Client.fPutObject("my-bucketname", "my-objectname", file, metaData, function (e) {
   if (e) {
     return console.log(e)
   }
@@ -43,8 +43,8 @@ s3Client.fPutObject('my-bucketname', 'my-objectname', file, metaData, function(e
 
 // Put a file in bucket my-bucketname with content-type detected automatically.
 // In this case it is `text/plain`.
-var file = 'my-testfile.txt'
-s3Client.fPutObject('my-bucketname', 'my-objectname', file, function(e) {
+var file = "my-testfile.txt"
+s3Client.fPutObject("my-bucketname", "my-objectname", file, function (e) {
   if (e) {
     return console.log(e)
   }
