@@ -17,45 +17,44 @@
 // Note: YOUR-ACCESSKEYID, YOUR-SECRETACCESSKEY, my-bucketname and my-objectname
 // are dummy values, please replace them with original values.
 
-
-var Minio = require('minio')
+var Minio = require("minio")
 
 var s3Client = new Minio.Client({
-  endPoint: 's3.amazonaws.com',
-  accessKey: 'YOUR-ACCESSKEYID',
-  secretKey: 'YOUR-SECRETACCESSKEY'
+  endPoint: "s3.amazonaws.com",
+  accessKey: "YOUR-ACCESSKEYID",
+  secretKey: "YOUR-SECRETACCESSKEY",
 })
 
 // Download the object my-objectname at an offset 1024, for a total of 4096 bytes.
 var size = 0
-s3Client.getPartialObject('my-bucketname', 'my-objectname', 1024, 4096, function(e, dataStream) {
+s3Client.getPartialObject("my-bucketname", "my-objectname", 1024, 4096, function (e, dataStream) {
   if (e) {
     return console.log(e)
   }
-  dataStream.on('data', function(chunk) {
+  dataStream.on("data", function (chunk) {
     size += chunk.length
   })
-  dataStream.on('end', function() {
+  dataStream.on("end", function () {
     console.log("End. Total size = " + size)
   })
-  dataStream.on('error', function(e) {
+  dataStream.on("error", function (e) {
     console.log(e)
   })
 })
 
 var versionedObjSize = 0
 // reads 30 bytes from the offset 10.
-s3Client.getPartialObject('mybucket', 'photo.jpg', 10, 30, {versionId:"my-versionId"},function(err, dataStream) {
+s3Client.getPartialObject("mybucket", "photo.jpg", 10, 30, { versionId: "my-versionId" }, function (err, dataStream) {
   if (err) {
     return console.log(err)
   }
-  dataStream.on('data', function(chunk) {
+  dataStream.on("data", function (chunk) {
     versionedObjSize += chunk.length
   })
-  dataStream.on('end', function() {
-    console.log('End. Total size = ' + versionedObjSize)
+  dataStream.on("end", function () {
+    console.log("End. Total size = " + versionedObjSize)
   })
-  dataStream.on('error', function(err) {
+  dataStream.on("error", function (err) {
     console.log(err)
   })
 })
