@@ -17,15 +17,15 @@
 // Note: YOUR-ACCESSKEYID, YOUR-SECRETACCESSKEY and my-bucketname are
 // dummy values, please replace them with original values.
 
-var Minio = require('minio')
+var Minio = require("minio")
 
 var s3Client = new Minio.Client({
-  endPoint: 's3.amazonaws.com',
-  accessKey: 'YOUR-ACCESSKEYID',
-  secretKey: 'YOUR-SECRETACCESSKEY'
+  endPoint: "s3.amazonaws.com",
+  accessKey: "YOUR-ACCESSKEYID",
+  secretKey: "YOUR-SECRETACCESSKEY",
 })
 // Create a bucket name my-bucketname.
-s3Client.makeBucket('my-bucketname', 'us-west-1', function(e) {
+s3Client.makeBucket("my-bucketname", "us-west-1", function (e) {
   if (e) {
     return console.log(e)
   }
@@ -33,7 +33,7 @@ s3Client.makeBucket('my-bucketname', 'us-west-1', function(e) {
 })
 
 // Create a bucket with object locking enabled.
-s3Client.makeBucket('mybucket', 'us-east-1', { ObjectLocking:true }, function(err) {
-  if (err) return console.log('Error creating bucket with lock .', err)
+s3Client.makeBucket("mybucket", "us-east-1", { ObjectLocking: true }, function (err) {
+  if (err) return console.log("Error creating bucket with lock .", err)
   console.log('Bucket created successfully in "us-east-1" and enabled object lock')
 })
