@@ -15,29 +15,36 @@
  */
 
 // Note: YOUR-ACCESSKEYID, YOUR-SECRETACCESSKEY, my-bucketname and my-objectname
-// are dummy values, please replace them with original values. 
+// are dummy values, please replace them with original values.
 
-var Minio = require('minio')
+var Minio = require("minio")
 
 var s3Client = new Minio.Client({
-  endPoint: 's3.amazonaws.com',
-  accessKey: 'YOUR-ACCESSKEYID',
-  secretKey: 'YOUR-SECRETACCESSKEY',
-  useSSL: true // Default is true.
+  endPoint: "s3.amazonaws.com",
+  accessKey: "YOUR-ACCESSKEYID",
+  secretKey: "YOUR-SECRETACCESSKEY",
+  useSSL: true, // Default is true.
 })
 
 // Presigned get object URL for my-objectname at my-bucketname, it expires in 7 days by default.
 var requestDate = new Date()
-requestDate.setHours(0,0,0,0)
+requestDate.setHours(0, 0, 0, 0)
 
-requestDate.setHours(0,0,0,0)
-s3Client.presignedGetObject('my-bucketname', 'my-objectname', 1000, {}, requestDate, function(e, presignedUrl) {
+requestDate.setHours(0, 0, 0, 0)
+s3Client.presignedGetObject("my-bucketname", "my-objectname", 1000, {}, requestDate, function (e, presignedUrl) {
   if (e) return console.log(e)
   console.log(presignedUrl)
 })
 
 // Versioning support
-s3Client.presignedGetObject('my-bucketname', 'my-objectname', 1000, {versionId:"10fa9946-3f64-4137-a58f-888065c0732e"},requestDate, function(e, presignedUrl) {
-  if (e) return console.log(e)
-  console.log(presignedUrl)
-})
+s3Client.presignedGetObject(
+  "my-bucketname",
+  "my-objectname",
+  1000,
+  { versionId: "10fa9946-3f64-4137-a58f-888065c0732e" },
+  requestDate,
+  function (e, presignedUrl) {
+    if (e) return console.log(e)
+    console.log(presignedUrl)
+  }
+)
