@@ -1,4 +1,3 @@
-
 /*
  * MinIO Javascript Library for Amazon S3 Compatible Cloud Storage, (C) 2021 MinIO, Inc.
  *
@@ -18,15 +17,15 @@
 // Note: YOUR-ACCESSKEYID, YOUR-SECRETACCESSKEY and my-bucketname are
 // dummy values, please replace them with original values.
 
-var Minio = require('minio')
+var Minio = require("minio")
 
 var s3Client = new Minio.Client({
-  endPoint: 's3.amazonaws.com',
-  accessKey: 'YOUR-ACCESSKEYID',
-  secretKey: 'YOUR-SECRETACCESSKEY'
+  endPoint: "s3.amazonaws.com",
+  accessKey: "YOUR-ACCESSKEYID",
+  secretKey: "YOUR-SECRETACCESSKEY",
 })
 
-s3Client.removeObjectTagging('bucketname', "object-name", function (err){
+s3Client.removeObjectTagging("bucketname", "object-name", function (err) {
   if (err) {
     return console.log(err)
   }
@@ -34,7 +33,7 @@ s3Client.removeObjectTagging('bucketname', "object-name", function (err){
 })
 
 //remove tags on a version of an object
-s3Client.removeObjectTagging('bucketname', 'object-name', { versionId: "my-object-version-id" }, function (err){
+s3Client.removeObjectTagging("bucketname", "object-name", { versionId: "my-object-version-id" }, function (err) {
   if (err) {
     return console.log(err)
   }
