@@ -14,19 +14,18 @@
  * limitations under the License.
  */
 
-
 // Note: YOUR-ACCESSKEYID, YOUR-SECRETACCESSKEY, my-bucketname and
 // my-objectname are dummy values, please replace them with original values.
 
-var Minio = require('minio')
+var Minio = require("minio")
 
 var s3Client = new Minio.Client({
-  endPoint: 's3.amazonaws.com',
-  accessKey: 'YOUR-ACCESSKEYID',
-  secretKey: 'YOUR-SECRETACCESSKEY'
+  endPoint: "s3.amazonaws.com",
+  accessKey: "YOUR-ACCESSKEYID",
+  secretKey: "YOUR-SECRETACCESSKEY",
 })
 // Remove an object name my-objectname.
-s3Client.removeObject('my-bucketname', 'my-objectname', function(e) {
+s3Client.removeObject("my-bucketname", "my-objectname", function (e) {
   if (e) {
     return console.log(e)
   }
@@ -34,7 +33,7 @@ s3Client.removeObject('my-bucketname', 'my-objectname', function(e) {
 })
 
 // Remove an object with name 'my-objectname' and a versionId.
-s3Client.removeObject('my-bucketname', 'my-objectname', {versionId:"my-versionId"}, function(e) {
+s3Client.removeObject("my-bucketname", "my-objectname", { versionId: "my-versionId" }, function (e) {
   if (e) {
     return console.log(e)
   }
@@ -42,15 +41,20 @@ s3Client.removeObject('my-bucketname', 'my-objectname', {versionId:"my-versionId
 })
 
 // Remove an object with name 'my-objectname' and a versionId with object renetion override with governanceBypass:true.
-s3Client.removeObject('my-bucketname', 'my-objectname', {versionId:"my-versionId", governanceBypass:true}, function(e) {
-  if (e) {
-    return console.log(e)
+s3Client.removeObject(
+  "my-bucketname",
+  "my-objectname",
+  { versionId: "my-versionId", governanceBypass: true },
+  function (e) {
+    if (e) {
+      return console.log(e)
+    }
+    console.log("Success")
   }
-  console.log("Success")
-})
+)
 
 // force delete object/prefix
-s3Client.removeObject('force-del-test', 'test/', {forceDelete:true}, function(e) {
+s3Client.removeObject("force-del-test", "test/", { forceDelete: true }, function (e) {
   if (e) {
     return console.log(e)
   }
