@@ -17,12 +17,12 @@
 // Note: YOUR-ACCESSKEYID, YOUR-SECRETACCESSKEY, and my-bucketname
 // are dummy values, please replace them with original values.
 
-var Minio = require('minio')
+var Minio = require("minio")
 
 var s3Client = new Minio.Client({
-  endPoint: 's3.amazonaws.com',
-  accessKey: 'YOUR-ACCESSKEYID',
-  secretKey: 'YOUR-SECRETACCESSKEY'
+  endPoint: "s3.amazonaws.com",
+  accessKey: "YOUR-ACCESSKEYID",
+  secretKey: "YOUR-SECRETACCESSKEY",
 })
 
 // Bucket policy - GET requests on "testbucket" bucket will not need authentication.
@@ -65,8 +65,8 @@ var policy = `
 }
 `
 
-s3Client.setBucketPolicy('testbucket', policy, (err) => {
-	if (err) throw err
+s3Client.setBucketPolicy("testbucket", policy, (err) => {
+  if (err) throw err
 
-	console.log('Set bucket policy')
+  console.log("Set bucket policy")
 })
