@@ -17,26 +17,31 @@
 // Note: YOUR-ACCESSKEYID, YOUR-SECRETACCESSKEY and my-bucketname are
 // dummy values, please replace them with original values.
 
-var Minio = require('minio')
+var Minio = require("minio")
 
 var s3Client = new Minio.Client({
-  endPoint: 's3.amazonaws.com',
-  accessKey: 'YOUR-ACCESSKEYID',
-  secretKey: 'YOUR-SECRETACCESSKEY'
+  endPoint: "s3.amazonaws.com",
+  accessKey: "YOUR-ACCESSKEYID",
+  secretKey: "YOUR-SECRETACCESSKEY",
 })
 
 //Set legal hold config of an object.
-s3Client.setObjectLegalHold('bucketName', 'objectName', {status:"ON"}, function(err, res) {
+s3Client.setObjectLegalHold("bucketName", "objectName", { status: "ON" }, function (err, res) {
   if (err) {
-    return console.log('Unable to set legal hold config for the object', err)
+    return console.log("Unable to set legal hold config for the object", err)
   }
-  console.log('Success')
+  console.log("Success")
 })
 
 //Set legal hold config of an object with versionId.
-s3Client.setObjectLegalHold('bucketName', 'objectName', { status:"ON", versionId:'my-obj-version-uuid' }, function(err, res) {
-  if (err) {
-    return console.log('Unable to set legal hold config for the object version', err)
+s3Client.setObjectLegalHold(
+  "bucketName",
+  "objectName",
+  { status: "ON", versionId: "my-obj-version-uuid" },
+  function (err, res) {
+    if (err) {
+      return console.log("Unable to set legal hold config for the object version", err)
+    }
+    console.log("Success")
   }
-  console.log('Success')
-})
+)
