@@ -17,16 +17,16 @@
 // Note: YOUR-ACCESSKEYID, YOUR-SECRETACCESSKEY and my-bucketname are
 // dummy values, please replace them with original values.
 
-var Minio = require('minio')
+var Minio = require("minio")
 
 var s3Client = new Minio.Client({
-  endPoint: 's3.amazonaws.com',
-  accessKey: 'YOUR-ACCESSKEYID',
-  secretKey: 'YOUR-SECRETACCESSKEY'
+  endPoint: "s3.amazonaws.com",
+  accessKey: "YOUR-ACCESSKEYID",
+  secretKey: "YOUR-SECRETACCESSKEY",
 })
 
 //Set Object lock on a bucket
-s3Client.setObjectLockConfig('my-bucketname', {mode:"COMPLIANCE",unit:'Days', validity:10 }, function (err){
+s3Client.setObjectLockConfig("my-bucketname", { mode: "COMPLIANCE", unit: "Days", validity: 10 }, function (err) {
   if (err) {
     return console.log(err)
   }
@@ -34,7 +34,7 @@ s3Client.setObjectLockConfig('my-bucketname', {mode:"COMPLIANCE",unit:'Days', va
 })
 
 //To reset/remove object lock config.
-s3Client.setObjectLockConfig('my-bucketname', {}, function (err){
+s3Client.setObjectLockConfig("my-bucketname", {}, function (err) {
   if (err) {
     return console.log(err)
   }
