@@ -14,19 +14,18 @@
  * limitations under the License.
  */
 
-
 // Note: YOUR-ACCESSKEYID, YOUR-SECRETACCESSKEY, my-bucketname
 // and my-objectname are dummy values, please replace them with original values.
 
-var Minio = require('minio')
+var Minio = require("minio")
 
 var s3Client = new Minio.Client({
-  endPoint: 's3.amazonaws.com',
-  accessKey: 'YOUR-ACCESSKEYID',
-  secretKey: 'YOUR-SECRETACCESSKEY'
+  endPoint: "s3.amazonaws.com",
+  accessKey: "YOUR-ACCESSKEYID",
+  secretKey: "YOUR-SECRETACCESSKEY",
 })
 // Get stat information for my-objectname.
-s3Client.statObject('my-bucketname', 'my-objectname', function(e, stat) {
+s3Client.statObject("my-bucketname", "my-objectname", function (e, stat) {
   if (e) {
     return console.log(e)
   }
@@ -35,7 +34,7 @@ s3Client.statObject('my-bucketname', 'my-objectname', function(e, stat) {
 
 // Get stat information for a specific version of 'my-objectname'
 //Bucket must be versioning enabled.
-s3Client.statObject('my-bucketname', 'my-objectname', {versionId:"my-uuid"},function(e, stat) {
+s3Client.statObject("my-bucketname", "my-objectname", { versionId: "my-uuid" }, function (e, stat) {
   if (e) {
     return console.log(e)
   }
