@@ -1,15 +1,11 @@
 import Credentials from "./Credentials"
 
 class CredentialProvider {
-  constructor({
-    accessKey,
-    secretKey,
-    sessionToken
-  }) {
+  constructor({ accessKey, secretKey, sessionToken }) {
     this.credentials = new Credentials({
       accessKey,
       secretKey,
-      sessionToken
+      sessionToken,
     })
   }
 
@@ -24,7 +20,6 @@ class CredentialProvider {
       throw new Error("Unable to set Credentials . it should be an instance of Credentials class")
     }
   }
-
 
   setAccessKey(accessKey) {
     this.credentials.setAccessKey(accessKey)
@@ -49,8 +44,6 @@ class CredentialProvider {
   getSessionToken() {
     return this.credentials.getSessionToken()
   }
-
-
 }
 
 export default CredentialProvider
