@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-import { EventEmitter } from 'events'
-import * as transformers from './transformers'
-import {DEFAULT_REGION, pipesetup, uriEscape} from './helpers'
+import { EventEmitter } from "events"
+
+import { DEFAULT_REGION, pipesetup, uriEscape } from "./helpers"
+import * as transformers from "./transformers"
 
 // Notification config - array of target configs.
 // Target configs can be
@@ -25,15 +26,15 @@ import {DEFAULT_REGION, pipesetup, uriEscape} from './helpers'
 // 3. CloudFront (lambda function)
 export class NotificationConfig {
   add(target) {
-    let instance = ''
+    let instance = ""
     if (target instanceof TopicConfig) {
-      instance = 'TopicConfiguration'
+      instance = "TopicConfiguration"
     }
     if (target instanceof QueueConfig) {
-      instance = 'QueueConfiguration'
+      instance = "QueueConfiguration"
     }
     if (target instanceof CloudFunctionConfig) {
-      instance = 'CloudFunctionConfiguration'
+      instance = "CloudFunctionConfiguration"
     }
     if (!this[instance]) this[instance] = []
     this[instance].push(target)
@@ -45,17 +46,17 @@ class TargetConfig {
   setId(id) {
     this.Id = id
   }
-  addEvent(newevent){
+  addEvent(newevent) {
     if (!this.Event) this.Event = []
     this.Event.push(newevent)
   }
   addFilterSuffix(suffix) {
-    if (!this.Filter) this.Filter = {S3Key : {FilterRule:[]}}
-    this.Filter.S3Key.FilterRule.push({Name:"suffix", Value:suffix})
+    if (!this.Filter) this.Filter = { S3Key: { FilterRule: [] } }
+    this.Filter.S3Key.FilterRule.push({ Name: "suffix", Value: suffix })
   }
   addFilterPrefix(prefix) {
-    if (!this.Filter) this.Filter = {S3Key : {FilterRule:[]}}
-    this.Filter.S3Key.FilterRule.push({Name:"prefix", Value:prefix})
+    if (!this.Filter) this.Filter = { S3Key: { FilterRule: [] } }
+    this.Filter.S3Key.FilterRule.push({ Name: "prefix", Value: prefix })
   }
 }
 
@@ -87,16 +88,15 @@ export const buildARN = (partition, service, region, accountId, resource) => {
   return "arn:" + partition + ":" + service + ":" + region + ":" + accountId + ":" + resource
 }
 
-
-export const ObjectCreatedAll                      = "s3:ObjectCreated:*"
-export const ObjectCreatedPut                      = "s3:ObjectCreated:Put"
-export const ObjectCreatedPost                     = "s3:ObjectCreated:Post"
-export const ObjectCreatedCopy                     = "s3:ObjectCreated:Copy"
-export const ObjectCreatedCompleteMultipartUpload  = "s3:ObjectCreated:CompleteMultipartUpload"
-export const ObjectRemovedAll                      = "s3:ObjectRemoved:*"
-export const ObjectRemovedDelete                   = "s3:ObjectRemoved:Delete"
-export const ObjectRemovedDeleteMarkerCreated      = "s3:ObjectRemoved:DeleteMarkerCreated"
-export const ObjectReducedRedundancyLostObject     = "s3:ReducedRedundancyLostObject"
+export const ObjectCreatedAll = "s3:ObjectCreated:*"
+export const ObjectCreatedPut = "s3:ObjectCreated:Put"
+export const ObjectCreatedPost = "s3:ObjectCreated:Post"
+export const ObjectCreatedCopy = "s3:ObjectCreated:Copy"
+export const ObjectCreatedCompleteMultipartUpload = "s3:ObjectCreated:CompleteMultipartUpload"
+export const ObjectRemovedAll = "s3:ObjectRemoved:*"
+export const ObjectRemovedDelete = "s3:ObjectRemoved:Delete"
+export const ObjectRemovedDeleteMarkerCreated = "s3:ObjectRemoved:DeleteMarkerCreated"
+export const ObjectReducedRedundancyLostObject = "s3:ReducedRedundancyLostObject"
 
 // Poll for notifications, used in #listenBucketNotification.
 // Listening constitutes repeatedly requesting s3 whether or not any
@@ -132,7 +132,7 @@ export class NotificationPoller extends EventEmitter {
     // Don't continue if we're looping again but are cancelled.
     if (this.ending) return
 
-    let method = 'GET'
+    let method = "GET"
     var queries = []
     if (this.prefix) {
       var prefix = uriEscape(this.prefix)
@@ -143,21 +143,21 @@ export class NotificationPoller extends EventEmitter {
       queries.push(`suffix=${suffix}`)
     }
     if (this.events) {
-      this.events.forEach(s3event => queries.push('events='+uriEscape(s3event)))
+      this.events.forEach((s3event) => queries.push("events=" + uriEscape(s3event)))
     }
     queries.sort()
 
-    var query = ''
+    var query = ""
     if (queries.length > 0) {
-      query = `${queries.join('&')}`
+      query = `${queries.join("&")}`
     }
     const region = this.client.region || DEFAULT_REGION
-    this.client.makeRequest({ method, bucketName: this.bucketName, query }, '', [200], region, true, (e, response) => {
-      if (e) return this.emit('error', e)
+    this.client.makeRequest({ method, bucketName: this.bucketName, query }, "", [200], region, true, (e, response) => {
+      if (e) return this.emit("error", e)
 
       let transformer = transformers.getNotificationTransformer()
       pipesetup(response, transformer)
-        .on('data', result => {
+        .on("data", (result) => {
           // Data is flushed periodically (every 5 seconds), so we should
           // handle it after flushing from the JSON parser.
           let records = result.Records
@@ -165,15 +165,15 @@ export class NotificationPoller extends EventEmitter {
           if (!records) records = []
 
           // Iterate over the notifications and emit them individually.
-          records.forEach(record => {
-            this.emit('notification', record)
+          records.forEach((record) => {
+            this.emit("notification", record)
           })
 
           // If we're done, stop.
           if (this.ending) response.destroy()
         })
-        .on('error', e => this.emit('error', e))
-        .on('end', () => {
+        .on("error", (e) => this.emit("error", e))
+        .on("end", () => {
           // Do it again, if we haven't cancelled yet.
           process.nextTick(() => {
             this.checkForChanges()
@@ -181,5 +181,4 @@ export class NotificationPoller extends EventEmitter {
         })
     })
   }
-
 }
