@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-import { Transform } from 'stream'
-import Crypto from 'crypto'
-import * as querystring from 'query-string'
+import Crypto from "crypto"
+import * as querystring from "query-string"
+import { Transform } from "stream"
+
 import { getVersionId, sanitizeETag } from "./helpers"
 
 // We extend Transform because Writable does not implement ._flush().
@@ -52,22 +53,22 @@ export default class ObjectUploader extends Transform {
     this.id = null
 
     // Handle errors.
-    this.on('error', err => {
+    this.on("error", (err) => {
       callback(err)
     })
   }
 
   _transform(chunk, encoding, callback) {
     this.emptyStream = false
-    let method = 'PUT'
-    let headers = {'Content-Length': chunk.length}
-    let md5digest = ''
+    let method = "PUT"
+    let headers = { "Content-Length": chunk.length }
+    let md5digest = ""
 
     // Calculate and set Content-MD5 header if SHA256 is not set.
     // This will happen only when there is a secure connection to the s3 server.
     if (!this.client.enableSHA256) {
-      md5digest = Crypto.createHash('md5').update(chunk).digest()
-      headers['Content-MD5'] = md5digest.toString('base64')
+      md5digest = Crypto.createHash("md5").update(chunk).digest()
+      headers["Content-MD5"] = md5digest.toString("base64")
     }
     // We can flush the object in one packet if it fits in one chunk. This is true
     // if the chunk size is smaller than the part size, signifying the end of the
@@ -78,19 +79,19 @@ export default class ObjectUploader extends Transform {
         method,
         // Set user metadata as this is not a multipart upload
         headers: Object.assign({}, this.metaData, headers),
-        query: '',
+        query: "",
         bucketName: this.bucketName,
-        objectName: this.objectName
+        objectName: this.objectName,
       }
 
-      this.client.makeRequest(options, chunk, [200], '', true, (err, response) => {
+      this.client.makeRequest(options, chunk, [200], "", true, (err, response) => {
         if (err) return callback(err)
         let result = {
           etag: sanitizeETag(response.headers.etag),
-          versionId :getVersionId(response.headers)
+          versionId: getVersionId(response.headers),
         }
         // Ignore the 'data' event so that the stream closes. (nodejs stream requirement)
-        response.on('data', () => {})
+        response.on("data", () => {})
 
         // Give the etag back, we're done!
 
@@ -109,13 +110,13 @@ export default class ObjectUploader extends Transform {
     // if it hasn't already been done. The write will be buffered until the upload has been
     // initiated.
     if (this.id === null) {
-      this.once('ready', () => {
+      this.once("ready", () => {
         this._transform(chunk, encoding, callback)
       })
 
       // Check for an incomplete previous upload.
       this.client.findUploadId(this.bucketName, this.objectName, (err, id) => {
-        if (err) return this.emit('error', err)
+        if (err) return this.emit("error", err)
 
         // If no upload ID exists, initiate a new one.
         if (!id) {
@@ -125,7 +126,7 @@ export default class ObjectUploader extends Transform {
             this.id = id
 
             // We are now ready to accept new chunks — this will flush the buffered chunk.
-            this.emit('ready')
+            this.emit("ready")
           })
 
           return
@@ -135,20 +136,20 @@ export default class ObjectUploader extends Transform {
 
         // Retrieve the pre-uploaded parts, if we need to resume the upload.
         this.client.listParts(this.bucketName, this.objectName, id, (err, etags) => {
-          if (err) return this.emit('error', err)
+          if (err) return this.emit("error", err)
 
           // It is possible for no parts to be already uploaded.
           if (!etags) etags = []
 
           // oldParts will become an object, allowing oldParts[partNumber].etag
-          this.oldParts = etags.reduce(function(prev, item) {
+          this.oldParts = etags.reduce(function (prev, item) {
             if (!prev[item.part]) {
               prev[item.part] = item
             }
             return prev
           }, {})
 
-          this.emit('ready')
+          this.emit("ready")
         })
       })
 
@@ -164,13 +165,13 @@ export default class ObjectUploader extends Transform {
       let oldPart = this.oldParts[partNumber]
 
       // Calulcate the md5 hash, if it has not already been calculated.
-      if(!md5digest) {
-        md5digest = Crypto.createHash('md5').update(chunk).digest()
+      if (!md5digest) {
+        md5digest = Crypto.createHash("md5").update(chunk).digest()
       }
 
-      if (oldPart && md5digest.toString('hex') === oldPart.etag) {
+      if (oldPart && md5digest.toString("hex") === oldPart.etag) {
         // The md5 matches, the chunk has already been uploaded.
-        this.etags.push({part: partNumber, etag: oldPart.etag})
+        this.etags.push({ part: partNumber, etag: oldPart.etag })
 
         callback()
         return
@@ -180,27 +181,28 @@ export default class ObjectUploader extends Transform {
     // Write the chunk with an uploader.
     let query = querystring.stringify({
       partNumber: partNumber,
-      uploadId: this.id
+      uploadId: this.id,
     })
 
     let options = {
-      method, query, headers,
+      method,
+      query,
+      headers,
       bucketName: this.bucketName,
-      objectName: this.objectName
+      objectName: this.objectName,
     }
 
-    this.client.makeRequest(options, chunk, [200], '', true, (err, response) => {
+    this.client.makeRequest(options, chunk, [200], "", true, (err, response) => {
       if (err) return callback(err)
 
       // In order to aggregate the parts together, we need to collect the etags.
       let etag = response.headers.etag
-      if (etag)
-        etag = etag.replace(/^"/, '').replace(/"$/, '')
+      if (etag) etag = etag.replace(/^"/, "").replace(/"$/, "")
 
-      this.etags.push({part: partNumber, etag})
+      this.etags.push({ part: partNumber, etag })
 
       // Ignore the 'data' event so that the stream closes. (nodejs stream requirement)
-      response.on('data', () => {})
+      response.on("data", () => {})
 
       // We're ready for the next chunk.
       callback()
@@ -209,25 +211,26 @@ export default class ObjectUploader extends Transform {
 
   _flush(callback) {
     if (this.emptyStream) {
-      let method = 'PUT'
-      let headers = Object.assign({}, this.metaData, {'Content-Length': 0})
+      let method = "PUT"
+      let headers = Object.assign({}, this.metaData, { "Content-Length": 0 })
       let options = {
-        method, headers,
-        query: '',
+        method,
+        headers,
+        query: "",
         bucketName: this.bucketName,
-        objectName: this.objectName
+        objectName: this.objectName,
       }
 
-      this.client.makeRequest(options, '', [200], '', true, (err, response) => {
+      this.client.makeRequest(options, "", [200], "", true, (err, response) => {
         if (err) return callback(err)
 
         let result = {
           etag: sanitizeETag(response.headers.etag),
-          versionId: getVersionId(response.headers)
+          versionId: getVersionId(response.headers),
         }
 
         // Ignore the 'data' event so that the stream closes. (nodejs stream requirement)
-        response.on('data', () => {})
+        response.on("data", () => {})
 
         // Give the etag back, we're done!
         process.nextTick(() => {
@@ -247,17 +250,16 @@ export default class ObjectUploader extends Transform {
 
     // This is called when all of the chunks uploaded successfully, thus
     // completing the multipart upload.
-    this.client.completeMultipartUpload(this.bucketName, this.objectName, this.id,
-                                        this.etags, (err, etag) => {
-                                          if (err) return callback(err)
+    this.client.completeMultipartUpload(this.bucketName, this.objectName, this.id, this.etags, (err, etag) => {
+      if (err) return callback(err)
 
-                                          // Call our callback on the next tick to allow the streams infrastructure
-                                          // to finish what its doing before we continue.
-                                          process.nextTick(() => {
-                                            this.callback(null, etag)
-                                          })
+      // Call our callback on the next tick to allow the streams infrastructure
+      // to finish what its doing before we continue.
+      process.nextTick(() => {
+        this.callback(null, etag)
+      })
 
-                                          callback()
-                                        })
+      callback()
+    })
   }
 }
