@@ -14,25 +14,25 @@
  * limitations under the License.
  */
 
-import { isString } from './helpers.js'
+import { isString } from "./helpers"
 
 // List of currently supported endpoints.
 let awsS3Endpoint = {
-  'us-east-1': 's3.amazonaws.com',
-  'us-east-2': 's3-us-east-2.amazonaws.com',
-  'us-west-1': 's3-us-west-1.amazonaws.com',
-  'us-west-2': 's3-us-west-2.amazonaws.com',
-  'ca-central-1': 's3.ca-central-1.amazonaws.com',
-  'eu-west-1': 's3-eu-west-1.amazonaws.com',
-  'eu-west-2': 's3-eu-west-2.amazonaws.com',
-  'sa-east-1': 's3-sa-east-1.amazonaws.com',
-  'eu-central-1': 's3-eu-central-1.amazonaws.com',
-  'ap-south-1': 's3-ap-south-1.amazonaws.com',
-  'ap-southeast-1': 's3-ap-southeast-1.amazonaws.com',
-  'ap-southeast-2': 's3-ap-southeast-2.amazonaws.com',
-  'ap-northeast-1': 's3-ap-northeast-1.amazonaws.com',
-  'cn-north-1': 's3.cn-north-1.amazonaws.com.cn',
-  'ap-east-1': 's3.ap-east-1.amazonaws.com'
+  "us-east-1": "s3.amazonaws.com",
+  "us-east-2": "s3-us-east-2.amazonaws.com",
+  "us-west-1": "s3-us-west-1.amazonaws.com",
+  "us-west-2": "s3-us-west-2.amazonaws.com",
+  "ca-central-1": "s3.ca-central-1.amazonaws.com",
+  "eu-west-1": "s3-eu-west-1.amazonaws.com",
+  "eu-west-2": "s3-eu-west-2.amazonaws.com",
+  "sa-east-1": "s3-sa-east-1.amazonaws.com",
+  "eu-central-1": "s3-eu-central-1.amazonaws.com",
+  "ap-south-1": "s3-ap-south-1.amazonaws.com",
+  "ap-southeast-1": "s3-ap-southeast-1.amazonaws.com",
+  "ap-southeast-2": "s3-ap-southeast-2.amazonaws.com",
+  "ap-northeast-1": "s3-ap-northeast-1.amazonaws.com",
+  "cn-north-1": "s3.cn-north-1.amazonaws.com.cn",
+  "ap-east-1": "s3.ap-east-1.amazonaws.com",
   // Add new endpoints here.
 }
 
@@ -45,5 +45,5 @@ export function getS3Endpoint(region) {
   if (endpoint) {
     return endpoint
   }
-  return 's3.amazonaws.com'
+  return "s3.amazonaws.com"
 }
