@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-import Crypto from 'crypto'
-import _ from 'lodash'
-import { uriEscape, getScope, isString, isObject, isArray, isNumber,
-  makeDateShort, makeDateLong } from './helpers.js'
-import * as errors from './errors.js'
+import Crypto from "crypto"
+import _ from "lodash"
 
-const signV4Algorithm = 'AWS4-HMAC-SHA256'
+import * as errors from "./errors"
+import { getScope, isArray, isNumber, isObject, isString, makeDateLong, makeDateShort, uriEscape } from "./helpers"
+
+const signV4Algorithm = "AWS4-HMAC-SHA256"
 
 // getCanonicalRequest generate a canonical request of style.
 //
@@ -55,30 +55,30 @@ function getCanonicalRequest(method, path, headers, signedHeaders, hashedPayload
     return acc
   }, [])
 
-  const requestResource = path.split('?')[0]
-  let requestQuery = path.split('?')[1]
-  if (!requestQuery) requestQuery = ''
+  const requestResource = path.split("?")[0]
+  let requestQuery = path.split("?")[1]
+  if (!requestQuery) requestQuery = ""
 
   if (requestQuery) {
     requestQuery = requestQuery
-      .split('&')
+      .split("&")
       .sort()
-      .map(element => element.indexOf('=') === -1 ? element + '=' : element)
-      .join('&')
+      .map((element) => (element.indexOf("=") === -1 ? element + "=" : element))
+      .join("&")
   }
 
   const canonical = []
   canonical.push(method.toUpperCase())
   canonical.push(requestResource)
   canonical.push(requestQuery)
-  canonical.push(headersArray.join('\n') + '\n')
-  canonical.push(signedHeaders.join(';').toLowerCase())
+  canonical.push(headersArray.join("\n") + "\n")
+  canonical.push(signedHeaders.join(";").toLowerCase())
   canonical.push(hashedPayload)
-  return canonical.join('\n')
+  return canonical.join("\n")
 }
 
 // generate a credential string
-function getCredential(accessKey, region, requestDate,serviceName="s3") {
+function getCredential(accessKey, region, requestDate, serviceName = "s3") {
   if (!isString(accessKey)) {
     throw new TypeError('accessKey should be of type "string"')
   }
@@ -88,7 +88,7 @@ function getCredential(accessKey, region, requestDate,serviceName="s3") {
   if (!isObject(requestDate)) {
     throw new TypeError('requestDate should be of type "object"')
   }
-  return `${accessKey}/${getScope(region, requestDate,serviceName)}`
+  return `${accessKey}/${getScope(region, requestDate, serviceName)}`
 }
 
 // Returns signed headers array - alphabetically sorted
@@ -123,14 +123,14 @@ function getSignedHeaders(headers) {
   //
   //      Is skipped for obvious reasons
 
-  const ignoredHeaders = ['authorization', 'content-length', 'content-type', 'user-agent']
+  const ignoredHeaders = ["authorization", "content-length", "content-type", "user-agent"]
   return _.map(headers, (v, header) => header)
-    .filter(header => ignoredHeaders.indexOf(header) === -1)
+    .filter((header) => ignoredHeaders.indexOf(header) === -1)
     .sort()
 }
 
 // returns the key used for calculating signature
-function getSigningKey(date, region, secretKey,serviceName="s3") {
+function getSigningKey(date, region, secretKey, serviceName = "s3") {
   if (!isObject(date)) {
     throw new TypeError('date should be of type "object"')
   }
@@ -141,14 +141,16 @@ function getSigningKey(date, region, secretKey,serviceName="s3") {
     throw new TypeError('secretKey should be of type "string"')
   }
   const dateLine = makeDateShort(date)
-  let hmac1 = Crypto.createHmac('sha256', 'AWS4' + secretKey).update(dateLine).digest(),
-    hmac2 = Crypto.createHmac('sha256', hmac1).update(region).digest(),
-    hmac3 = Crypto.createHmac('sha256', hmac2).update(serviceName).digest()
-  return Crypto.createHmac('sha256', hmac3).update('aws4_request').digest()
+  let hmac1 = Crypto.createHmac("sha256", "AWS4" + secretKey)
+      .update(dateLine)
+      .digest(),
+    hmac2 = Crypto.createHmac("sha256", hmac1).update(region).digest(),
+    hmac3 = Crypto.createHmac("sha256", hmac2).update(serviceName).digest()
+  return Crypto.createHmac("sha256", hmac3).update("aws4_request").digest()
 }
 
 // returns the string that needs to be signed
-function getStringToSign(canonicalRequest, requestDate, region,serviceName="s3") {
+function getStringToSign(canonicalRequest, requestDate, region, serviceName = "s3") {
   if (!isString(canonicalRequest)) {
     throw new TypeError('canonicalRequest should be of type "string"')
   }
@@ -158,14 +160,14 @@ function getStringToSign(canonicalRequest, requestDate, region,serviceName="s3")
   if (!isString(region)) {
     throw new TypeError('region should be of type "string"')
   }
-  const hash = Crypto.createHash('sha256').update(canonicalRequest).digest('hex')
+  const hash = Crypto.createHash("sha256").update(canonicalRequest).digest("hex")
   const scope = getScope(region, requestDate, serviceName)
   const stringToSign = []
   stringToSign.push(signV4Algorithm)
   stringToSign.push(makeDateLong(requestDate))
   stringToSign.push(scope)
   stringToSign.push(hash)
-  const signString = stringToSign.join('\n')
+  const signString = stringToSign.join("\n")
   return signString
 }
 
@@ -184,11 +186,11 @@ export function postPresignSignatureV4(region, date, secretKey, policyBase64) {
     throw new TypeError('policyBase64 should be of type "string"')
   }
   const signingKey = getSigningKey(date, region, secretKey)
-  return Crypto.createHmac('sha256', signingKey).update(policyBase64).digest('hex').toLowerCase()
+  return Crypto.createHmac("sha256", signingKey).update(policyBase64).digest("hex").toLowerCase()
 }
 
 // Returns the authorization header
-export function signV4(request, accessKey, secretKey, region, requestDate, serviceName="s3") {
+export function signV4(request, accessKey, secretKey, region, requestDate, serviceName = "s3") {
   if (!isObject(request)) {
     throw new TypeError('request should be of type "object"')
   }
@@ -203,28 +205,29 @@ export function signV4(request, accessKey, secretKey, region, requestDate, servi
   }
 
   if (!accessKey) {
-    throw new errors.AccessKeyRequiredError('accessKey is required for signing')
+    throw new errors.AccessKeyRequiredError("accessKey is required for signing")
   }
   if (!secretKey) {
-    throw new errors.SecretKeyRequiredError('secretKey is required for signing')
+    throw new errors.SecretKeyRequiredError("secretKey is required for signing")
   }
 
-  const sha256sum = request.headers['x-amz-content-sha256']
+  const sha256sum = request.headers["x-amz-content-sha256"]
 
   const signedHeaders = getSignedHeaders(request.headers)
-  const canonicalRequest = getCanonicalRequest(request.method, request.path, request.headers,
-                                               signedHeaders, sha256sum)
+  const canonicalRequest = getCanonicalRequest(request.method, request.path, request.headers, signedHeaders, sha256sum)
   const serviceIdentifier = serviceName || "s3"
-  const stringToSign = getStringToSign(canonicalRequest, requestDate, region,serviceIdentifier)
-  const signingKey = getSigningKey(requestDate, region, secretKey,serviceIdentifier)
+  const stringToSign = getStringToSign(canonicalRequest, requestDate, region, serviceIdentifier)
+  const signingKey = getSigningKey(requestDate, region, secretKey, serviceIdentifier)
   const credential = getCredential(accessKey, region, requestDate, serviceIdentifier)
-  const signature = Crypto.createHmac('sha256', signingKey).update(stringToSign).digest('hex').toLowerCase()
+  const signature = Crypto.createHmac("sha256", signingKey).update(stringToSign).digest("hex").toLowerCase()
 
-  return `${signV4Algorithm} Credential=${credential}, SignedHeaders=${signedHeaders.join(';').toLowerCase()}, Signature=${signature}`
+  return `${signV4Algorithm} Credential=${credential}, SignedHeaders=${signedHeaders
+    .join(";")
+    .toLowerCase()}, Signature=${signature}`
 }
 
-export function signV4ByServiceName( request, accessKey, secretKey, region, requestDate, serviceName="s3") {
-  return signV4(request, accessKey, secretKey, region,requestDate, serviceName)
+export function signV4ByServiceName(request, accessKey, secretKey, region, requestDate, serviceName = "s3") {
+  return signV4(request, accessKey, secretKey, region, requestDate, serviceName)
 }
 // returns a presigned URL string
 export function presignSignatureV4(request, accessKey, secretKey, sessionToken, region, requestDate, expires) {
@@ -242,53 +245,52 @@ export function presignSignatureV4(request, accessKey, secretKey, sessionToken, 
   }
 
   if (!accessKey) {
-    throw new errors.AccessKeyRequiredError('accessKey is required for presigning')
+    throw new errors.AccessKeyRequiredError("accessKey is required for presigning")
   }
   if (!secretKey) {
-    throw new errors.SecretKeyRequiredError('secretKey is required for presigning')
+    throw new errors.SecretKeyRequiredError("secretKey is required for presigning")
   }
 
   if (!isNumber(expires)) {
     throw new TypeError('expires should be of type "number"')
   }
   if (expires < 1) {
-    throw new errors.ExpiresParamError('expires param cannot be less than 1 seconds')
+    throw new errors.ExpiresParamError("expires param cannot be less than 1 seconds")
   }
   if (expires > 604800) {
-    throw new errors.ExpiresParamError('expires param cannot be greater than 7 days')
+    throw new errors.ExpiresParamError("expires param cannot be greater than 7 days")
   }
 
   const iso8601Date = makeDateLong(requestDate)
   const signedHeaders = getSignedHeaders(request.headers)
   const credential = getCredential(accessKey, region, requestDate)
-  const hashedPayload = 'UNSIGNED-PAYLOAD'
+  const hashedPayload = "UNSIGNED-PAYLOAD"
 
   const requestQuery = []
   requestQuery.push(`X-Amz-Algorithm=${signV4Algorithm}`)
   requestQuery.push(`X-Amz-Credential=${uriEscape(credential)}`)
   requestQuery.push(`X-Amz-Date=${iso8601Date}`)
   requestQuery.push(`X-Amz-Expires=${expires}`)
-  requestQuery.push(`X-Amz-SignedHeaders=${uriEscape(signedHeaders.join(';').toLowerCase())}`)
+  requestQuery.push(`X-Amz-SignedHeaders=${uriEscape(signedHeaders.join(";").toLowerCase())}`)
   if (sessionToken) {
     requestQuery.push(`X-Amz-Security-Token=${uriEscape(sessionToken)}`)
   }
 
-  const resource = request.path.split('?')[0]
-  let query = request.path.split('?')[1]
+  const resource = request.path.split("?")[0]
+  let query = request.path.split("?")[1]
   if (query) {
-    query = query + '&' + requestQuery.join('&')
+    query = query + "&" + requestQuery.join("&")
   } else {
-    query = requestQuery.join('&')
+    query = requestQuery.join("&")
   }
 
-  const path = resource + '?' + query
+  const path = resource + "?" + query
 
-  const canonicalRequest = getCanonicalRequest(request.method, path,
-                                               request.headers, signedHeaders, hashedPayload)
+  const canonicalRequest = getCanonicalRequest(request.method, path, request.headers, signedHeaders, hashedPayload)
 
   const stringToSign = getStringToSign(canonicalRequest, requestDate, region)
   const signingKey = getSigningKey(requestDate, region, secretKey)
-  const signature = Crypto.createHmac('sha256', signingKey).update(stringToSign).digest('hex').toLowerCase()
-  const presignedUrl = request.protocol + '//' + request.headers.host + path + `&X-Amz-Signature=${signature}`
+  const signature = Crypto.createHmac("sha256", signingKey).update(stringToSign).digest("hex").toLowerCase()
+  const presignedUrl = request.protocol + "//" + request.headers.host + path + `&X-Amz-Signature=${signature}`
   return presignedUrl
 }
